requirements.txt
board==1.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Cython==0.29.24
dataclasses==0.8
filelock==3.0.12
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.1
joblib==1.0.1
numpy~=1.21.1
packaging==21.0
pandas==1.1.5
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
sacremoses==0.0.45
scipy==1.5.4
semantic-version==2.8.5
sentencepiece==0.1.96
setuptools-rust==0.12.1
six==1.16.0

thop~=0.0.31-2005241907
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchvision~=0.10.0
tqdm==4.61.2
transformers~=4.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
